feat(mcp-server): wire hybrid search into standalone MCP + extract search()
- Move the hybrid search pipeline (symbol + BM25 + vector + RRF + rerank)
  from packages/index/src/service.ts into packages/index/src/retrieval/search.ts
  as a single source of truth shared by IndexService and the standalone MCP.
- apps/mcp-server: code_search now calls the full pipeline instead of the
  symbol-only stub. index_status returns populated + blob/chunk/symbol/ref
  counts. New reindex tool exposes a full pass to agents.
- Smoke test asserts the expanded 7-tool surface, the new status fields,
  and exercises code_search end-to-end against a tiny indexed fixture.
